Update terraform-exec and terraform-json depedencies for 1.1.0 support
Terraform 1.1.0 has been released:
https://www.hashicorp.com/blog/terraform-1-1-improves-refactoring-and-the-cloud-cli-experience.

It seems to bring a new version of the state file (version 1.0), so the
dependencies of Rover that work with the state file need to be updated in order
to support this new version.

This updates both terraform-json and terraform-exec, as terraform-exec depends
on terraform-json so it also needs a version bump for it to use the latest
terraform-json version.